Added X86.Movd and X86.Pextrd instructions (untested)

+++ Source/Mosa.UnitTest.System/UnitTestSystem.cs
@@ -469,7 +469,7 @@ namespace Mosa.UnitTest.System
 			}
 			else if (parameter is UInt32)
 			{
-				cmd.Add((int)parameter);
+				cmd.Add((int)((uint)parameter));
 			}
 			else if (parameter is UInt64)
 			{
